Use different hook for action_mailer config
By using the `after: "finisher_hook"` we are 
alleviated from the specific load order of the
initializers. ActionMailer allows the addition
of delivery methods at any point during app
execution as long as it's there before you send
mail.

Now we can lose the specific initializer path
tests, the generated initializer is unneeded
in a standard application, and the app can use
any method to set the configuration during 
initialization and it's picked up at the end.